+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Add `willReadStringTree` callback argument to `Builder::build` and retire
+  `Builder::treesRead`
 * Update `Watcher` and `Builder` interaction to prevent double builds.
 
 # 0.11.0